backend/requirements-agent.txt
--pre
agent-framework
azure-identity>=1.17
python-dotenv>=1.0
tiktoken>=0.7
